requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge-score
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
python-box
ensure
fastapi
uvicorn
Jinja2

-e .